requirements.txt
asgiref==3.8.1
distlib==0.3.9
Django==5.2.1
django-cleanup==9.0.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
filelock==3.18.0
pillow==11.2.1
platformdirs==4.3.8
PyJWT==2.9.0
sqlparse==0.5.3
tzdata==2025.2
virtualenv==20.31.2
